added file operations

+++ my_device.c
@@ -5,8 +5,8 @@ MODULE_LICENSE("GPL v2");
 
 extern int open_d(struct inode *inode, struct file *filp);
 extern int release_d(struct inode *inode, struct file *filp);
-extern int read_from_d(struct file *filp, char __user *ubuf, size_t len, loff_t *offset);
-extern int write_to_d(struct file *filp, const char __user *ubuf, size_t len, loff_t *offset);
+extern ssize_t read_from_d(struct file *filp, char __user *ubuf, size_t len, loff_t *offset);
+extern ssize_t write_to_d(struct file *filp, const char __user *ubuf, size_t len, loff_t *offset);
 
 struct my_device_t my_device = {
     .fops = {
